- remove unused VendorServices.swift duplicate - factor out app-specific VendorServices into app, to allow calls to openURL() - add support for new Branding class setting `branding.send-feedback-url` with template support
