+++ ownCloudAppShared/Tools/VendorServices.swift
@@ -17,7 +17,6 @@
 */
 
 import UIKit
-import MessageUI
 import ownCloudSDK
 import ownCloudApp
 
@@ -125,66 +124,6 @@ public class VendorServices : NSObject {
 	}()
 
 	// MARK: - Vendor services
-	public func recommendToFriend(from viewController: UIViewController) {
-
-		guard let appStoreLink = self.classSetting(forOCClassSettingsKey: .appStoreLink) as? String else {
-				return
-		}
-		let appName = VendorServices.shared.appName
-
-		let message = """
-		<p>I want to invite you to use \(appName) on your smartphone!</p>
-		<a href="\(appStoreLink)">Download here</a>
-		"""
-		self.sendMail(to: nil, subject: "Try \(appName) on your smartphone!", message: message, from: viewController)
-	}
-
-	public func sendFeedback(from viewController: UIViewController) {
-		var buildType = "release".localized
-
-		if self.isBetaBuild {
-			buildType = "beta".localized
-		}
-
-		var appSuffix = ""
-		if OCLicenseEMMProvider.isEMMVersion {
-			appSuffix = "-EMM"
-		}
-
-		guard let feedbackEmail = self.feedbackMail else {
-			return
-		}
-		self.sendMail(to: feedbackEmail, subject: "\(self.appVersion) (\(self.appBuildNumber)) \(buildType) \(self.appName)\(appSuffix)", message: nil, from: viewController)
-	}
-
-	public func sendMail(to: String?, subject: String?, message: String?, from viewController: UIViewController) {
-		if MFMailComposeViewController.canSendMail() {
-			let mail = MFMailComposeViewController()
-			mail.mailComposeDelegate = self
-			if to != nil {
-				mail.setToRecipients([to!])
-			}
-
-			if subject != nil {
-				mail.setSubject(subject!)
-			}
-
-			if message != nil {
-				mail.setMessageBody(message!, isHTML: true)
-			}
-
-			viewController.present(mail, animated: true)
-		} else {
-			let alert = ThemedAlertController(title: "Please configure an email account".localized,
-											  message: "You need to configure an email account first to be able to send emails.".localized,
-											  preferredStyle: .alert)
-
-			let okAction = UIAlertAction(title: "OK", style: .default, handler: nil)
-			alert.addAction(okAction)
-			viewController.present(alert, animated: true)
-		}
-	}
-
 	public func considerReviewPrompt() {
 		guard
 			let reviewPromptEnabled = self.classSetting(forOCClassSettingsKey: .enableReviewPrompt) as? Bool,
@@ -213,12 +152,6 @@ public class VendorServices : NSObject {
 
 		userDefaults.setValue(true, forKey: UserDefaultsKeys.notFirstAppLaunch.rawValue)
 		userDefaults.synchronize()
-	}
-}
-
-extension VendorServices: MFMailComposeViewControllerDelegate {
-	public func mailComposeController(_ controller: MFMailComposeViewController, didFinishWith result: MFMailComposeResult, error: Error?) {
-		controller.dismiss(animated: true)
 	}
 }
